+++ resources/ReleaseNotes.md
@@ -5,3 +5,4 @@
 # Software
 
 - Rename `AutomatedAnalysis` to `IMAPTestingAnalysis`
+- Rename `Result` to `PSD`